feat: add BlueButton component with stylesfix some config eslint
