+++ poker/src/components/BlueButton/BlueButton.module.scss
@@ -1,0 +1,34 @@
+.BlueButton_blue,
+.BlueButton_white {
+  margin: 0 auto;
+  padding: 0 10px;
+  min-width: 240px;
+  border: 1px solid #496a81;
+  border-radius: 2px;
+  font-family: 'Ruda', sans-serif;
+  font-size: 24px;
+  line-height: 42px;
+  font-weight: bold;
+}
+.BlueButton_blue {
+  color: white;
+  background-color: #2b3a67;
+}
+.BlueButton_white {
+  color: #2b3a67;
+  background-color: white;
+}
+.BlueButton_white:hover {
+  color: white;
+  background-color: #2b3a67;
+}
+.BlueButton_blue:hover {
+  color: #2b3a67;
+  background-color: white;
+}
+.BlueButton_white:active {
+  filter: opacity(0.8);
+}
+.BlueButton_blue:active {
+  filter: opacity(0.8);
+}
